+++ pgmax/fg/nodes.py
@@ -1,7 +1,7 @@
-"""A module containing classes that specify the components of a Factor Graph."""
+"""A module containing classes that specify the basic components of a Factor Graph."""
 
 from dataclasses import asdict, dataclass
-from typing import Mapping, Tuple, Union
+from typing import Any, Mapping, Tuple, Union
 
 import jax
 import jax.numpy as jnp
@@ -26,7 +26,7 @@ class Variable:
 
 @jax.tree_util.register_pytree_node_class
 @dataclass(frozen=True, eq=False)
-class EnumerationWiring:
+class Wiring:
     """Wiring for enumeration factors.
 
     Args:
@@ -34,15 +34,10 @@ class EnumerationWiring:
             Number of states for the variables connected to each edge
         var_states_for_edges: Array of shape (num_edge_states,)
             Global variable state indices for each edge state
-        factor_configs_edge_states: Array of shape (num_factor_configs, 2)
-            factor_configs_edge_states[ii] contains a pair of global enumeration factor_config and global edge_state indices
-            factor_configs_edge_states[ii, 0] contains the global enumeration factor config index
-            factor_configs_edge_states[ii, 1] contains the corresponding global edge_state index
     """
 
     edges_num_states: Union[np.ndarray, jnp.ndarray]
     var_states_for_edges: Union[np.ndarray, jnp.ndarray]
-    factor_configs_edge_states: Union[np.ndarray, jnp.ndarray]
 
     def __post_init__(self):
         for field in self.__dataclass_fields__:
@@ -57,111 +52,15 @@ class EnumerationWiring:
         return cls(**aux_data.unflatten(children))
 
 
-@jax.tree_util.register_pytree_node_class
 @dataclass(frozen=True, eq=False)
-class LogicalWiring:
-    """Wiring for LogicalFactors.
-
-    Args:
-        edges_num_states: Array of shape (2 * num_parents,)
-            Number of states for the variables connected to each edge
-        var_states_for_edges:  Array of shape (2 * (num_parents + num_children),)
-            Global parent and children variable state indices for each edge state
-        parents_edge_states: Array of shape (num_parents, 2)
-            parents_edge_states[ii, 0] contains the global factor index,
-            which takes into account all the logical factors
-            parents_edge_states[ii, 1] contains the message index of the parent variable's state 0,
-            which takes into account all the enumeration and logical factors
-            The parent variable's state 1 is parents_edge_states[ii, 2] + 1
-        children_edge_states: Array of shape (num_factors,)
-            children_edge_states[ii] contains the message index of the child variable's state 0,
-            which takes into account all the enumeration and logical factors
-            The child variable's state 1 is children_edge_states[ii, 1] + 1
-    """
-
-    edges_num_states: Union[np.ndarray, jnp.ndarray]
-    var_states_for_edges: Union[np.ndarray, jnp.ndarray]
-    parents_edge_states: Union[np.ndarray, jnp.ndarray]
-    children_edge_states: Union[np.ndarray, jnp.ndarray]
-
-    def __post_init__(self):
-        for field in self.__dataclass_fields__:
-            if isinstance(getattr(self, field), np.ndarray):
-                getattr(self, field).flags.writeable = False
-
-    def tree_flatten(self):
-        return jax.tree_util.tree_flatten(asdict(self))
-
-    @classmethod
-    def tree_unflatten(cls, aux_data, children):
-        return cls(**aux_data.unflatten(children))
-
-
-@jax.tree_util.register_pytree_node_class
-@dataclass(frozen=True, eq=False)
-class ORWiring(LogicalWiring):
-    pass
-
-
-@dataclass(frozen=True, eq=False)
-class EnumerationFactor:
-    """An enumeration factor
+class Factor:
+    """A factor
 
     Args:
         variables: List of connected variables
-        configs: Array of shape (num_val_configs, num_variables)
-            An array containing an explicit enumeration of all valid configurations
-        log_potentials: Array of shape (num_val_configs,). An array containing
-            the log of the potential value for every possible configuration
-
-    Raises:
-        ValueError: If:
-            (1) The dtype of the configs array is not int
-            (2) The dtype of the potential array is not float
-            (3) Configs does not have the correct shape
-            (4) The potential array does not have the correct shape
-            (5) The configs array contains invalid values
     """
 
     variables: Tuple[Variable, ...]
-    configs: np.ndarray
-    log_potentials: np.ndarray
-
-    def __post_init__(self):
-        self.configs.flags.writeable = False
-        if not np.issubdtype(self.configs.dtype, np.integer):
-            raise ValueError(
-                f"Configurations should be integers. Got {self.configs.dtype}."
-            )
-
-        if not np.issubdtype(self.log_potentials.dtype, np.floating):
-            raise ValueError(
-                f"Potential should be floats. Got {self.log_potentials.dtype}."
-            )
-
-        if self.configs.ndim != 2:
-            raise ValueError(
-                "configs should be a 2D array containing a list of valid configurations for "
-                f"EnumerationFactor. Got a configs array of shape {self.configs.shape}."
-            )
-
-        if len(self.variables) != self.configs.shape[1]:
-            raise ValueError(
-                f"Number of variables {len(self.variables)} doesn't match given configurations {self.configs.shape}"
-            )
-
-        if self.log_potentials.shape != (self.configs.shape[0],):
-            raise ValueError(
-                f"Expected log potentials of shape {(self.configs.shape[0],)} for "
-                f"({self.configs.shape[0]}) valid configurations. Got log potentials of "
-                f"shape {self.log_potentials.shape}."
-            )
-
-        vars_num_states = np.array([variable.num_states for variable in self.variables])
-        if not np.logical_and(
-            self.configs >= 0, self.configs < vars_num_states[None]
-        ).all():
-            raise ValueError("Invalid configurations for given variables")
 
     @utils.cached_property
     def edges_num_states(self) -> np.ndarray:
@@ -176,32 +75,8 @@ class EnumerationFactor:
         )
         return edges_num_states
 
-    @utils.cached_property
-    def factor_configs_edge_states(self) -> np.ndarray:
-        """Array containing factor configs and edge states pairs
-
-        Returns:
-            Array of shape (num_factor_configs, 2)
-            factor_configs_edge_states[ii] contains a pair of global factor_config and edge_state indices
-            factor_configs_edge_states[ii, 0] contains the global factor config index,
-            which takes into account all the enumeration factors
-            factor_configs_edge_states[ii, 1] contains the corresponding global edge_state index,
-            which takes into account all the enumeration and logical factors
-        """
-        edges_starts = np.insert(self.edges_num_states.cumsum(), 0, 0)[:-1]
-        factor_configs_edge_states = np.stack(
-            [
-                np.repeat(np.arange(self.configs.shape[0]), self.configs.shape[1]),
-                (self.configs + edges_starts[None]).flatten(),
-            ],
-            axis=1,
-        )
-        return factor_configs_edge_states
-
-    def compile_wiring(
-        self, vars_to_starts: Mapping[Variable, int]
-    ) -> EnumerationWiring:
-        """Compile enumeration wiring for the enumeration factor
+    def compile_wiring(self, vars_to_starts: Mapping[Variable, int]) -> Any:
+        """Compile wiring for the factor
 
         Args:
             vars_to_starts: A dictionary that maps variables to their global starting indices
@@ -209,157 +84,8 @@ class EnumerationFactor:
                 of its n variable states are m, m + 1, ..., m + n - 1
 
         Returns:
-            Enumeration wiring for the enumeration factor
-        """
-        var_states_for_edges = np.concatenate(
-            [
-                np.arange(variable.num_states) + vars_to_starts[variable]
-                for variable in self.variables
-            ]
-        )
-        return EnumerationWiring(
-            edges_num_states=self.edges_num_states,
-            var_states_for_edges=var_states_for_edges,
-            factor_configs_edge_states=self.factor_configs_edge_states,
-        )
-
-
-@dataclass(frozen=True, eq=False)
-class LogicalFactor:
-    """A logical OR/AND factor
-
-    Args:
-        parents_variables: List of parents variables
-        child_variable: Child variable
-
-    Raises:
-        ValueError: If:
-            (1) The parents variables are not all binary
-            (2) The child variable is not binary
-            (3) The list of parent variables is empty
-    """
-
-    parents_variables: Tuple[Variable, ...]
-    child_variable: Variable
-    log_potentials: None
-
-    def __post_init__(self):
-        if not np.all(
-            [variable.num_states == 2 for variable in self.parents_variables]
-        ):
-            raise ValueError("Parents variables should all be binary")
-
-        if not self.child_variable.num_states == 2:
-            raise ValueError("Child variables should be binary")
-
-        if len(self.parents_variables) == 0:
-            raise ValueError("At least one parent variable is required")
-
-    @utils.cached_property
-    def edges_num_states(self) -> np.ndarray:
-        """Number of states for the variables connected to each edge
-
-        Returns:
-            Array of shape (num_edges,)
-            Number of states for the variables connected to each edge
-        """
-        return np.full((len(self.parents_variables) + 1,), fill_value=2, dtype=int)
-
-    @utils.cached_property
-    def variables(self) -> np.ndarray:
-        """
-        Returns:
-            All the variables of the logical factor
-        """
-        return self.parents_variables + (self.child_variable,)
-
-    @utils.cached_property
-    def parents_edge_states(self) -> np.ndarray:
-        """
-        Returns:
-            The parents variables edge states.
-        """
-        num_parents = len(self.parents_variables)
-
-        parents_edge_states = np.vstack(
-            [
-                np.zeros(num_parents, dtype=int),
-                np.arange(0, 2 * num_parents, 2, dtype=int),
-            ],
-        ).T
-        return parents_edge_states
-
-    @utils.cached_property
-    def child_edge_state(self) -> np.ndarray:
-        """
-        Returns:
-            The child variable edge states.
-        """
-        return np.array([2 * len(self.parents_variables)], dtype=int)
-
-    def get_var_states_for_edges(
-        self, vars_to_starts: Mapping[Variable, int]
-    ) -> np.ndarray:
-        """Compile the array of variable states for edges
-
-        Args:
-            vars_to_starts: A dictionary that maps variables to their global starting indices
-                For an n-state variable, a global start index of m means the global indices
-                of its n variable states are m, m + 1, ..., m + n - 1
-
-        Returns:
-            Variable states for edges array
-        """
-        parents_var_states_for_edges = np.concatenate(
-            [
-                np.arange(variable.num_states) + vars_to_starts[variable]
-                for variable in self.parents_variables
-            ]
-        )
-        child_var_states_for_edges = (
-            np.arange(self.child_variable.num_states)
-            + vars_to_starts[self.child_variable]
-        )
-
-        var_states_for_edges = np.concatenate(
-            [parents_var_states_for_edges, child_var_states_for_edges]
-        )
-        return var_states_for_edges
-
-    def compile_wiring(self, vars_to_starts: Mapping[Variable, int]) -> LogicalWiring:
-        """Compile logical wiring for the logical factor
-
-        Args:
-            vars_to_starts: A dictionary that maps variables to their global starting indices
-                For an n-state variable, a global start index of m means the global indices
-                of its n variable states are m, m + 1, ..., m + n - 1
-
-        Returns:
-            Logical wiring for the logical factor
+            a dictionary mapping all possible names to different variables.
         """
         raise NotImplementedError(
-            "Please subclass the LogicalFactor class and override this method"
-        )
-
-
-@dataclass(frozen=True, eq=False)
-class ORFactor(LogicalFactor):
-    def compile_wiring(self, vars_to_starts: Mapping[Variable, int]) -> ORWiring:
-        """Compile OR wiring for the OR factor
-
-        Args:
-            vars_to_starts: A dictionary that maps variables to their global starting indices
-                For an n-state variable, a global start index of m means the global indices
-                of its n variable states are m, m + 1, ..., m + n - 1
-
-        Returns:
-            OR wiring for the OR factor
-        """
-        var_states_for_edges = self.get_var_states_for_edges(vars_to_starts)
-
-        return ORWiring(
-            edges_num_states=self.edges_num_states,
-            var_states_for_edges=var_states_for_edges,
-            parents_edge_states=self.parents_edge_states,
-            children_edge_states=self.child_edge_state,
+            "Please subclass the VariableGroup class and override this method"
         )
